# src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.0'
  title: Care Platform
  description: 'This is the API definition for the Care Platform.

  Authentication  Nodes:
  Click Authorize to enter User and Component authorization codes.

  - userAuth: The JWT token provided by the login command. Since the swagger 2.0 (OpenAPI) specification does
       not natively support bearer authentication. Please use this APIKey field. Prefix it with the word Bearer,
       like this: Bearer JWT_TOKEN
  - componentAuth: The key of the component, this can be obtained by making GET requests to .../components.'
host: localhost:10010
basePath: /v0
schemes:
- http
#s
consumes:
- application/json
produces:
- application/json
securityDefinitions:
   userAuth:
     type: apiKey
     in: header
     name: X-API-Key

   componentAuth:
     type: apiKey
     in: header
     name: X-API-Key
paths:
  /users:
    x-swagger-router-controller: users
    post:
      description: Create a new user with only basic details which can be later associated with a client
      summary: Create a new user
      tags:
      - user
      operationId: createUser
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: New user details
        schema:
          $ref: '#/definitions/UserWithPassword'
      responses:
        201:
          description: Successfully created
        406:
          description: Data is invalid
          schema:
            $ref: '#/definitions/ValidationError'
  /users/me:
    x-swagger-router-controller: users
  
    get:
      security:
      - userAuth: []
      description: Get basic details of your own user including, name, email address and password.
      summary: Get your own details on the platform
      tags:
      - user
      operationId: getUser
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        401:
          description: You have not logged in or an invalid or expired token has been supplied

    put:
      security:
      - userAuth: []
      description: Update basic details of your own user including, name, email address and password. Information for fields which should not be updated should not  be supplied.
      summary: Update your own user account once authenticated
      tags:
      - user
      operationId: updateUser
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: New user details
        schema:
          $ref: '#/definitions/UserWithPassword'
      responses:
        200:
          description: Successfully updated
        401:
          description: You have not logged in or an invalid or expired token has been supplied
        406:
          description: The input is not valid
      
    delete:
      security:
      - userAuth: []
      description: Delete your own user account. If you are the last user associated with a client, that client will also be deleted. You will need to be authenticated and provide your password in the body of the request to confirm the deletion.
      summary: Delete your own user account once authenticated
      tags:
      - user
      operationId: deleteUser
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: Credentials Pair
        schema:
          $ref: '#/definitions/Credentials'
      responses:
        200:
          description: Successfully deleted
        401:
          description: You have not logged in or an invalid or expired token has been supplied
        404:
          description: User account no longer exists
        406:
          description: Invalid password supplied
  /users/login:
    x-swagger-router-controller: users
  
    post:
      description: Log in with your email address and password and obtain an  authorization token
      summary: User login
      tags:
      - user
      operationId: loginUser
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: Credentials Pair
        schema:
          $ref: '#/definitions/Credentials'
      responses:
        200:
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/JWT'
        401:
          description: Invalid credentials supplied
        404:
          description: User does not exist
  /users/passwordResetEmail:
    x-swagger-router-controller: users
  
    post:
      description: Specify email address to send a password reset email to the email  address of a registered user.
      summary: Send a password recovery email
      tags:
      - user
      operationId: sendPasswordResetEmail
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: Email address for an exisiting user
        schema:
          $ref: '#/definitions/UsersPasswordresetemailRequest'
      responses:
        200:
          description: Email Sent
        404:
          description: User does not exist
  /users/passwordReset:
    x-swagger-router-controller: users
    post:
      description: Using your email address, and temporary password reset code, provide a new password for your account.
      summary: Change your password using a temporary password reset code
      tags:
      - user
      operationId: usePasswordResetCode
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: New user details
        schema:
          $ref: '#/definitions/PasswordReset'
      responses:
        200:
          description: User account updated
        401:
          description: Invalid password reset code
        404:
          description: User does not exist
        406:
          description: New password is invalid
...
definitions:
...
  UserWithPassword:
    title: UserWithPassword
    type: object
    properties:
      name:
        $ref: '#/definitions/Name'
      email:
        example: carer@example.com
        type: string
      password:
        example: password
        type: string
  Name:
    title: Name
    type: object
    properties:
      title:
        example: Dr
        type: string
      firstNames:
        example: Grace
        type: string
      surnames:
        example: Hopper
        type: string
  User:
    title: User
    type: object
    properties:
      name:
        $ref: '#/definitions/Name'
      email:
        example: carer@example.com
        type: string
  Credentials:
    title: Credentials
    type: object
    properties:
      email:
        example: carer@example.com
        type: string
      password:
        example: password
        type: string
    required:
    - email
    - password
  JWT:
    title: JWT
    type: object
    properties:
      token:
        description: JSON Web Token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ.eyJlbWFpbCI6Im9sYXR1bmRlZ2FydWJhQGdtYWlsLmNvbSIsImZ1bGxOYW1lIjoiT2xhdHVuZGUgR2FydWJhIiwiX2lkIjoiNThmMjYzNDdiMTY1YzUxODM1NDMxYTNkIiwiaW F0IjoxNDkyMjgwMTk4fQ.VcMpybz08cB5PsrMSr25En4_EwCGWZVFgciO4M3ENE
        type: string
  UsersPasswordresetemailRequest:
    title: Users PasswordResetEmail Request
    type: object
    properties:
      email:
        example: carer@example.com
        type: string
  PasswordReset:
    title: PasswordReset
    type: object
    properties:
      email:
        example: carer@example.com
        type: string
      code:
        example: SA92JS9a029384JASKJN
        type: string
      newPassword:
        example: password
        type: string
    required:
    - email
    - code
    - newPassword
...
  ValidationError:
   title: Validation Error
   type: string
   example: "User already exists"